Ensure new fields can be reactively added to table
Instead of using the index of a field in the fields array to control
its visibility, use a dictionary of key / isVisible pairs.
This enables the code to detect if a field has been added reactively
and therefore set its default visibility correctly.

Note that this change now means that field key values must be unique
for column visibility control to work correctly (this wasn’t the case
before).  If there are non-unique key names, the visibility of the
first such field in the fields array will govern the visibility of all
of those fields.

This commit also includes corrected and added package tests related to
this change, and note in the README about consequences of non-unique
field keys.

Fixes #156